Move Rhythm classes to their own files
