+++ wheatley/rhythm/__init__.py
@@ -1,0 +1,5 @@
+""" A module to hold various classes determining when Wheatley rings his bells. """
+
+from .wait_for_user import WaitForUserRhythm
+from .regression import RegressionRhythm
+from .abstract_rhythm import Rhythm
